add enrollment histories endpoints

+++ app/Http/Controllers/StudentController.php
@@ -20,7 +20,14 @@ class StudentController extends Controller
         $start_date = $request->start_date ? Carbon::createFromFormat('Y-m-d H:i:s', $request->start_date)->startOfDay()->toDateTimeString() : null;
         $end_date = $request->end_date ? Carbon::createFromFormat('Y-m-d H:i:s', $request->end_date)->endOfDay()->toDateTimeString() : null;
         $like = 'LIKE';
-        $paginated = Student::withTrashed()->with(['level', 'program', 'registration.educationBackgrounds', 'registration.guardians', 'registration.level', 'registration.program', 'registration.schoolSetting', 'registration.siblings', 'schoolSetting'])
+        $paginated = Student::withTrashed()->with([
+            'level', 'program', 'registration.educationBackgrounds',
+            'registration.guardians', 'registration.level', 'registration.program',
+            'registration.schoolSetting', 'registration.siblings', 'schoolSetting', 'enrollmentHistories' => function ($query) {
+                $query->with(['schoolSetting.term', 'enrolledSubjects'])->leftJoin('school_settings', 'school_setting_id', '=', 'school_settings.id')
+                    ->orderBy('school_settings.academic_year')->orderBy('school_settings.term_id')->select('enrollment_histories.*');
+            }
+        ])->withCount(['enrollmentHistories'])
             ->when($start_date && $end_date, function ($query) use ($start_date, $end_date) {
                 $query->whereHas(
                     'registration',
@@ -30,19 +37,21 @@ class StudentController extends Controller
                 );
             })
             ->when($search, function ($query) use ($like, $search) {
-                $query->whereHas(
-                    'registration',
-                    function ($query) use ($like, $search) {
-                        $query->where(
-                            function ($query) use ($like, $search) {
-                                $query->where('last_name', $like,  '%' . $search . '%')
-                                    ->orWhere('first_name', $like,  '%' . $search . '%');
-                            }
-                        );
-                    }
-                )->orWhere('student_number', $like, '%' . $search . '%');
+                $query->where(function ($query) use ($like, $search) {
+                    $query->whereHas(
+                        'registration',
+                        function ($query) use ($like, $search) {
+                            $query->where(
+                                function ($query) use ($like, $search) {
+                                    $query->where('last_name', $like,  '%' . $search . '%')
+                                        ->orWhere('first_name', $like,  '%' . $search . '%');
+                                }
+                            );
+                        }
+                    )->orWhere('student_number', $like, '%' . $search . '%');
+                });
             })
-            ->when($request->student_status != 'ALL', function ($query) use ($request) {
+            ->when($request->exists('student_status') && $request->student_status != 'ALL', function ($query) use ($request) {
                 $query->when($request->student_status == 'ACTIVE', function ($query) {
                     $query->whereNull('deleted_at');
                 })
@@ -50,13 +59,13 @@ class StudentController extends Controller
                         $query->whereNotNull('deleted_at');
                     });
             })
-            ->when(boolval($request->p), function ($query) use ($request) {
+            ->when($request->exists('p'), function ($query) use ($request) {
                 $query->whereIn('program_id', $request->input('p'));
             })
-            ->when(boolval($request->l), function ($query) use ($request) {
+            ->when($request->exists('l'), function ($query) use ($request) {
                 $query->whereIn('level_id', $request->input('l'));
             })
-            ->when(boolval($request->st), function ($query) use ($request) {
+            ->when($request->exists('st'), function ($query) use ($request) {
                 $query->whereIn('student_type_id', $request->input('st'));
             })
             ->paginate($limit);
@@ -77,24 +86,31 @@ class StudentController extends Controller
     /**
      * Display the specified resource.
      *
-     * @param  \App\Models\Student  $student
+     * @param  number  $student
      * @return \Illuminate\Http\Response
      */
-    public function show(Student $student)
+    public function show($student)
     {
-        //
+        return Student::withTrashed()->with([
+            'level', 'program', 'registration.educationBackgrounds', 'registration.guardians',
+            'registration.level', 'registration.program', 'registration.schoolSetting',
+            'registration.siblings', 'schoolSetting', 'enrollmentHistories' => function ($query) {
+                $query->with(['schoolSetting.term', 'enrolledSubjects'])->leftJoin('school_settings', 'school_setting_id', '=', 'school_settings.id')
+                    ->orderBy('school_settings.academic_year')->orderBy('school_settings.term_id')->select('enrollment_histories.*');
+            }
+        ])->withCount(['enrollmentHistories'])->findOrFail($student);
     }
 
     /**
      * Update the specified resource in storage.
      *
      * @param  \Illuminate\Http\Request  $request
-     * @param  \App\Models\Student  $student
+     * @param  number  $student
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, Student $student)
+    public function update(Request $request, $student)
     {
-        //
+        $updated_student = Student::with('registration')->findOrFail($student);
     }
 
     /**
@@ -106,10 +122,22 @@ class StudentController extends Controller
      */
     public function destroy(Request $request, $student)
     {
-        $student = Student::withTrashed()->findOrFail($student);
+        $student = Student::withTrashed()->withCount(['enrollmentHistories'])->findOrFail($student);
 
         if (boolval($request->forceDelete)) {
+            if ($student->enrollment_histories_count > 0) {
+                return response()->json(
+                    [
+                        "error" => "Unable to Delete Student!",
+                        "message" => "Student is currently associated to $student->enrollment_histories_count enrollments",
+                    ],
+                    500
+                );
+            }
+
+            $student->user->roles()->detach();
             $student->forceDelete();
+            $student->user()->forceDelete();
             return response()->json(
                 [
                     "message" => "Student Records Permanently Deleted!",
@@ -120,8 +148,10 @@ class StudentController extends Controller
 
         if (boolval($request->toggle)) {
             if ($student->trashed()) {
+                $student->user()->restore();
                 $student->restore();
             } else {
+                $student->user->delete();
                 $student->delete();
             }
             return response()->json(
